Update word controller with respecting proper authentication

+++ src/main/java/dev/mlukas/langreader/text/WordController.java
@@ -1,13 +1,16 @@
 package dev.mlukas.langreader.text;
 
 import dev.mlukas.langreader.language.Language;
+import dev.mlukas.langreader.language.NoChosenLanguageException;
 import dev.mlukas.langreader.user.User;
 import dev.mlukas.langreader.user.UserService;
+import org.checkerframework.checker.nullness.qual.Nullable;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
+import java.security.Principal;
 
 @RestController
 @RequestMapping("/words")
@@ -22,11 +25,15 @@ public class WordController {
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void addWord(@Valid @RequestBody TokenUpdateRequest token) {
-        User foundUser = userService.getUser(UserService.MARTIN);
-        Language chosenLang = foundUser.getChosenLang();
-        String value = token.value();
+    public void addWord(@Valid @RequestBody TokenUpdateRequest token, Principal principal) {
+        User foundUser = userService.getUser(principal.getName());
 
+        @Nullable Language chosenLang = foundUser.getChosenLang();
+        if (chosenLang == null) {
+            throw new NoChosenLanguageException(foundUser.getUsername());
+        }
+
+        String value = token.value();
         if (wordService.existBy(value, chosenLang, foundUser)) {
             throw new DuplicateWordException(
                     "The word '%s' is already present in your database. For updating its type, use PUT method.".formatted(value)
@@ -39,13 +46,18 @@ public class WordController {
 
     @PutMapping
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updateWord(@Valid @RequestBody TokenUpdateRequest tokenRequest) {
-        User foundUser = userService.getUser(UserService.MARTIN);
+    public void updateWord(@Valid @RequestBody TokenUpdateRequest tokenRequest, Principal principal) {
+        User foundUser = userService.getUser(principal.getName());
+
+        @Nullable Language chosenLang = foundUser.getChosenLang();
+        if (chosenLang == null) {
+            throw new NoChosenLanguageException(foundUser.getUsername());
+        }
 
         String value = tokenRequest.value().toLowerCase();
         WordType newType = tokenRequest.type();
         if (tokenRequest.type() != WordType.UNKNOWN) { // Instead of setting unknown type, just delete the word
-            Word foundWord = wordService.getWordBy(value, foundUser.getChosenLang(), foundUser);
+            Word foundWord = wordService.getWordBy(value, chosenLang, foundUser);
             if (foundWord.getType() != newType) {
                 foundWord.setType(newType);
                 wordService.save(foundWord);
@@ -55,8 +67,14 @@ public class WordController {
 
     @DeleteMapping
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteWord(@NotBlank @RequestParam("word") String value) {
-        User foundUser = userService.getUser(UserService.MARTIN);
+    public void deleteWord(@NotBlank @RequestParam("word") String value, Principal principal) {
+        User foundUser = userService.getUser(principal.getName());
+
+        @Nullable Language chosenLang = foundUser.getChosenLang();
+        if (chosenLang == null) {
+            throw new NoChosenLanguageException(foundUser.getUsername());
+        }
+
         Word foundWord = wordService.getWordBy(value.toLowerCase(), foundUser.getChosenLang(), foundUser);
         wordService.delete(foundWord);
     }
